Refactor resume menu files
