+++ src/menu/resume_helpers.py
@@ -1,112 +1,15 @@
 """
-src/menu/resume.py
+src/menu/resume_helpers.py
 
-Menu option for creating and viewing resume snapshots.
+Helper functions for building and rendering resume snapshots.
 """
-
 import json
-from datetime import datetime
 from typing import List, Dict, Any
 
-from src.db import (
-    get_all_user_project_summaries,
-    insert_resume_snapshot,
-    list_resumes,
-    get_resume_snapshot,
-)
 from src.models.project_summary import ProjectSummary
-from src.insights.rank_projects.rank_project_importance import collect_project_data
 
 
-def view_resume_items(conn, user_id: int, username: str):
-    """
-    Prompt to create a new resume snapshot or view an existing one.
-    """
-    while True:
-        print("\nResume options:")
-        print("1. Create a new resume from current projects")
-        print("2. View an existing resume snapshot")
-        print("3. Back to main menu")
-        choice = input("Select an option (1-3): ").strip()
-
-        if choice == "1":
-            _handle_create_resume(conn, user_id, username)
-            return
-        elif choice == "2":
-            handled = _handle_view_existing_resume(conn, user_id)
-            if handled:
-                return
-            # otherwise loop back to resume menu
-        elif choice == "3":
-            return
-        else:
-            print("Invalid choice, please enter 1, 2, or 3.")
-
-
-def _handle_create_resume(conn, user_id: int, username: str):
-    summaries = _load_project_summaries(conn, user_id)
-    if not summaries:
-        print("No project summaries available. Run an analysis first.")
-        return
-
-    # Rank projects and select top 5 by score
-    ranked = collect_project_data(conn, user_id)
-    top_names = [name for name, _score in ranked[:5]]
-    if top_names:
-        summaries = [s for s in summaries if s.project_name in top_names]
-        print(f"[Resume] Using top {len(summaries)} ranked projects: {', '.join(top_names)}")
-
-    snapshot = _build_resume_snapshot(summaries)
-    rendered = _render_snapshot(snapshot)
-
-    default_name = f"Resume {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}"
-    name = input(f"Enter a name for this resume [{default_name}]: ").strip() or default_name
-
-    resume_json = json.dumps(snapshot, default=str)
-    insert_resume_snapshot(conn, user_id, name, resume_json, rendered)
-    print(f"\n[Resume] Saved snapshot '{name}'.")
-
-
-def _handle_view_existing_resume(conn, user_id: int) -> bool:
-    resumes = list_resumes(conn, user_id)
-    if not resumes:
-        print("No saved resumes yet. Create one first.")
-        return False
-
-    print("\nSaved resumes:")
-    for idx, r in enumerate(resumes, 1):
-        print(f"{idx}. {r['name']} (created {r['created_at']})")
-
-    choice = input("Select a resume to view (number) or press Enter to cancel: ").strip()
-    if not choice.isdigit():
-        print("Cancelled.")
-        return False
-
-    idx = int(choice)
-    if idx < 1 or idx > len(resumes):
-        print("Invalid selection.")
-        return False
-
-    resume_id = resumes[idx - 1]["id"]
-    record = get_resume_snapshot(conn, user_id, resume_id)
-    if not record:
-        print("Unable to load the selected resume.")
-        return False
-
-    # Prefer stored rendered text; fall back to rendering from stored JSON.
-    if record.get("rendered_text"):
-        print("\n" + record["rendered_text"])
-    else:
-        try:
-            snapshot = json.loads(record["resume_json"])
-            _render_snapshot(snapshot, print_output=True)
-        except Exception:
-            print("Stored resume is corrupted or unreadable.")
-            return False
-
-    return True
-
-def _load_project_summaries(conn, user_id: int) -> List[ProjectSummary]:
+def load_project_summaries(conn, user_id: int, get_all_user_project_summaries) -> List[ProjectSummary]:
     """Fetch and deserialize all saved project summaries for the user."""
     rows = get_all_user_project_summaries(conn, user_id)
     projects: List[ProjectSummary] = []
@@ -122,7 +25,7 @@ def _load_project_summaries(conn, user_id: int) -> List[ProjectSummary]:
     return projects
 
 
-def _build_resume_snapshot(summaries: List[ProjectSummary]) -> Dict[str, Any]:
+def build_resume_snapshot(summaries: List[ProjectSummary]) -> Dict[str, Any]:
     """Build a structured snapshot of resume data from project summaries."""
     projects = []
     for ps in summaries:
@@ -156,7 +59,7 @@ def _build_resume_snapshot(summaries: List[ProjectSummary]) -> Dict[str, Any]:
     return {"projects": projects, "aggregated_skills": agg}
 
 
-def _render_snapshot(snapshot: Dict[str, Any], print_output: bool = True) -> str:
+def render_snapshot(snapshot: Dict[str, Any], print_output: bool = True) -> str:
     """Render a snapshot to text; optionally print to console."""
     lines: List[str] = []
